# pkg/app/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
      trace_id:
        example: 1234567890abcdef1234567890abcdef
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.Tenant:
    properties:
      description:
        maxLength: 255
        type: string
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      status:
        enum:
        - active
        - blocked
        type: string
    required:
    - hash
    - status
    type: object
...
  protected.tenants:
    properties:
      tenants:
        items:
          $ref: '#/definitions/models.Tenant'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    email: admin@vxcontrol.com
    name: Dmitry Nagibin
    url: https://vxcontrol.com
  description: Swagger API for VXControl SOLDR backend product.
  license:
    name: Proprietary License
    url: https://raw.githubusercontent.com/vxcontrol/soldr/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SOLDR Swagger API
  version: "1.0"
paths:
...
  /tenant/{hash}:
    put:
      parameters:
      - description: tenant model to update
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      - description: tenant hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: tenant updated successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Tenant'
              type: object
        "400":
          description: invalid tenant request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating tenant not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: tenant not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating tenant
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update tenant
      tags:
      - Tenants
  /tenants/:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"..."}
            field value should be integer or string or array type
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: en
        description: Language of result data
        enum:
        - en
        - ru
        in: query
        name: lang
        required: true
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 100, -1 means unlimited)
        in: query
        maximum: 100
        minimum: -1
        name: pageSize
        required: true
        type: integer
      - default: '{}'
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
        in: query
        name: sort
        required: true
        type: string
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: tenants list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/protected.tenants'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting tenants not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting tenants
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve tenants list
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: tenant model to create from
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "201":
          description: tenant created successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Tenant'
              type: object
        "400":
          description: invalid tenant request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: creating tenant not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on creating tenant
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new tenant
      tags:
      - Tenants
  /tenants/{hash}:
    delete:
      parameters:
      - description: hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: tenant deleted successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: deleting tenant not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: tenant not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on deleting tenant
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete tenant by hash
      tags:
      - Tenants
    get:
      parameters:
      - description: hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: tenant received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Tenant'
              type: object
        "403":
          description: getting tenant not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: tenant not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting tenant
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve tenant by hash
      tags:
      - Tenants
...
swagger: "2.0"